feat: create addon-operator package
